[chore] add titles to edit dialog +...
 - get rid of empty helpers
 - rename profile_helper to backoffice_helper
 - bring the common_editor back to where it belongs
