+++ app/views/achievements/edit.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "achievements/form", profile: nil, achievement: @achievement
 } %>
