+++ app/views/achievements/new.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "achievements/form", profile: @profile, achievement: @achievement
 } %>
