+++ app/views/educations/new.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "educations/form", profile: @profile, education: @education
 } %>
