+++ app/views/experiences/edit.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "experiences/form", profile: nil, experience: @experience
 } %>
