+++ app/views/experiences/new.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "experiences/form", profile: @profile, experience: @experience
 } %>
