+++ app/views/infosciences/edit.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "infosciences/form", profile: nil, infoscience: @infoscience
 } %>
