+++ app/views/infosciences/new.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "infosciences/form", profile: @profile, infoscience: @infoscience
 } %>
