+++ app/views/publications/edit.turbo_stream.erb
@@ -1,3 +1,3 @@
-<%= common_editor {
+<%= common_editor(title: t(".form_title")) {
   render "publications/form", profile: nil, publication: @publication
 } %>
