+++ config/locales/achievements/de.yml
@@ -11,6 +11,10 @@ de:
     placeholder:
       achievement: {}
   achievements:
+    edit:
+      form_title: ''
+    new:
+      form_title: ''
     form:
       category: ''
       description: ''
