+++ config/locales/achievements/fr.yml
@@ -11,6 +11,10 @@ fr:
     placeholder:
       achievement: {}
   achievements:
+    edit:
+      form_title: "Editing Achievement"
+    new:
+      form_title: "Creating a new Achievement"
     form:
       category: Catégorie
       description: Description
