+++ config/locales/achievements/it.yml
@@ -11,6 +11,10 @@ it:
     placeholder:
       achievement: {}
   achievements:
+    edit:
+      form_title: ''
+    new:
+      form_title: ''
     form:
       category: ''
       description: ''
