+++ config/locales/awards/de.yml
@@ -21,5 +21,9 @@ de:
         url: url
         year: Jahr
   awards:
+    edit:
+      form_title: ''
+    new:
+      form_title: ''
     form:
       help: {}
