+++ config/locales/awards/it.yml
@@ -21,5 +21,9 @@ it:
         url: url
         year: Year
   awards:
+    edit:
+      form_title: ''
+    new:
+      form_title: ''
     form:
       help: {}
