+++ config/locales/educations/fr.yml
@@ -27,5 +27,9 @@ fr:
         year_begin: de (année)
         year_end: à (année)
   educations:
+    edit:
+      form_title: "Editing Education"
+    new:
+      form_title: "Creating a new Education"
     form:
       period: période (en années)
