+++ config/locales/educations/it.yml
@@ -26,5 +26,9 @@ it:
         year_begin: from year
         year_end: to year
   educations:
+    edit:
+      form_title: ''
+    new:
+      form_title: ''
     form:
       period: period (years)
