+++ config/locales/infosciences/en.yml
@@ -21,5 +21,9 @@ en:
         title: Title
         url: infoscience-exports url
   infosciences:
+    edit:
+      form_title: "Editing Infoscience export link"
+    new:
+      form_title: "Creating Infoscience export link"
     form:
       create: Save
