+++ config/locales/socials/en.yml
@@ -26,6 +26,10 @@ en:
   socials:
     add:
       add: Add Social Media Link or Research ID
+    edit:
+      form_title: "Editing Research ID"
+    new:
+      form_title: "Creating a new Research ID"
     form:
       identifier_label:
         facebook: Facebook ID
